Fixed tick overflow error when recording the same action for over 255 ticks

+++ bot_go_brr/src/record.rs
@@ -5,7 +5,7 @@ use alloc::string::ToString;
 
 pub struct Record {
     arg: DriveArg,
-    held: u8,
+    held: u16,
 }
 
 impl Record {
